ClutterInputDevice: Store the cursor coordinate state as floating point
To support sub-pixel motion events coming from relative events, the
fraction part needs to be stored in the input device state as well. To
do this, simply change the current type from gint to gfloat.

https://bugzilla.gnome.org/show_bug.cgi?id=736413

+++ clutter/clutter-device-manager-private.h
@@ -65,8 +65,8 @@ typedef struct _ClutterTouchInfo
   ClutterEventSequence *sequence;
   ClutterActor *actor;
 
-  gint current_x;
-  gint current_y;
+  gfloat current_x;
+  gfloat current_y;
 } ClutterTouchInfo;
 
 struct _ClutterInputDevice
@@ -106,8 +106,8 @@ struct _ClutterInputDevice
   ClutterStage *stage;
 
   /* the current state */
-  gint current_x;
-  gint current_y;
+  gfloat current_x;
+  gfloat current_y;
   guint32 current_time;
   gint current_button_number;
   ClutterModifierType current_state;
@@ -161,8 +161,8 @@ void            _clutter_input_device_remove_event_sequence     (ClutterInputDev
                                                                  ClutterEvent         *event);
 void            _clutter_input_device_set_coords                (ClutterInputDevice   *device,
                                                                  ClutterEventSequence *sequence,
-                                                                 gint                  x,
-                                                                 gint                  y,
+                                                                 gfloat                x,
+                                                                 gfloat                y,
                                                                  ClutterStage         *stage);
 void            _clutter_input_device_set_state                 (ClutterInputDevice   *device,
                                                                  ClutterModifierType   state);
